annoation burn - simplification

+++ src/main/java/uk/gov/hmcts/reform/em/npa/service/dto/external/annotation/RectangleDTO.java
@@ -1,7 +1,5 @@
 package uk.gov.hmcts.reform.em.npa.service.dto.external.annotation;
 
-import uk.gov.hmcts.reform.em.npa.service.dto.AbstractAuditingDTO;
-
 import java.io.Serializable;
 import java.util.Objects;
 import java.util.UUID;
@@ -13,13 +11,13 @@ public class RectangleDTO extends AbstractAuditingDTO implements Serializable {
 
     private UUID id;
 
-    private Float x;
+    private Double x;
 
-    private Float y;
+    private Double y;
 
-    private Float width;
+    private Double width;
 
-    private Float height;
+    private Double height;
 
     private UUID annotationId;
 
@@ -31,35 +29,35 @@ public class RectangleDTO extends AbstractAuditingDTO implements Serializable {
         this.id = id;
     }
 
-    public Float getX() {
+    public Double getX() {
         return x;
     }
 
-    public void setX(Float x) {
+    public void setX(Double x) {
         this.x = x;
     }
 
-    public Float getY() {
+    public Double getY() {
         return y;
     }
 
-    public void setY(Float y) {
+    public void setY(Double y) {
         this.y = y;
     }
 
-    public Float getWidth() {
+    public Double getWidth() {
         return width;
     }
 
-    public void setWidth(Float width) {
+    public void setWidth(Double width) {
         this.width = width;
     }
 
-    public Float getHeight() {
+    public Double getHeight() {
         return height;
     }
 
-    public void setHeight(Float height) {
+    public void setHeight(Double height) {
         this.height = height;
     }
